Remove `event_handler` module (gnosis/gp-v2-services#1486)
As a follow up to #1485, this PR removes the `event_handler` module and the `BalancerPoolRegistry` type as it is no longer needed (it was superseded by the `Aggregate: InternalPoolFetching` type from the aforementioned PR). We still index events, it was just mostly moved to `pool_storage` and `pool_fetching::registry` modules so that is generic on the factory type.

The logic in `pool_fetching` is unchanged (despite the PR changes). It would just moved into smaller helper methods to make the test setup and assertions a little easier.

### Test Plan

There was a very important unit test in the `event_handler` module that would index all pools by events and compare them to what the Balancer subgraph would return. This unit test has been re-written and now will actually fetch pool state for all those pools (on top of the requirement of indexing the events).

```
$ cargo test -p shared -- balancer_fetched_pools_match_subgraph --ignored --nocapture
...
2021-12-15T21:47:20.368Z  WARN shared::sources::balancer_v2::pool_fetching::tests: unknown_pools=[]
test sources::balancer_v2::pool_fetching::tests::balancer_fetched_pools_match_subgraph ... ok
...
```