Validate Price Estimation Queries Early and at a Single Point (gnosis/gp-v2-services#1499)
* Add wrapper for estimators which sanitizes queries

This wrapper takes an impl PriceEstimating and handles trivial price
estimations before passing it onto the wrapped estimator in case it is
really needed.
This was done in order to avoid duplicating the logic for those trivial
price estimation cases within every price estimator.

* Wrap CompetitionPriceEstimator in SanitizedPriceEstimator

This will cause trivial and invalid price estimation Queries to be
handled immediately instead of passing them to each and every price
estimator.
Also price estimations for ETH will be modified to be price estimations
for our native token instead.

* Drop bad token detection from price estimators

Because every price estimation query now goes through the
SanitizedPriceEstimator which detects unsupported tokens, the code
responsible for that can be removed from the different existing price
estimators.
This code was also the only reason why the estimators stored the
bad_token_detector so I removed that as well.

* Detect unsupported tokens upfront

We can build a cache of unsupported tokens upfront to avoid network
calls for tokens where we already know whether they are supported or not
because we already queried them at some point.

* Make use of bulk price estimations

Bulk price estimations are usually at least as efficient as individual
estimations. To take advantage of that, trivial cases get solved
immediately by the SanitizedPriceEstimator while "difficult" Queries
get collected in a Vec which gets forwarded to the bulk estimation API
of the wrapped estimator.
After the "difficult" Queries got estimated, they get merged back into
the trivial cases while preserving the expected order.

* Remove unwrap from detecting unsupported tokens

The change to detect unsupported tokens in bulk complicated the error
handling a bit.
Instead of a simple set which includes tokens with errors, we map an
unsupported token to an error. This can be an "UnsupportedToken" error
from a succesful bad token detection or an "Other" error (most likely
related to networking issues).
With the HashMap it's way relatively painless to propagate the correct
error for each failed price estimation query.

* Assert all forwarded estimations have been merged

After merge_trivial_and_forwarded_estimations is done all
AwaitingEthEstimation and AwaitingErc20Estimation have to replaced with
a forwarded result.
We also expect PriceEstimating::estimates() to give us exactly as many
results as we have queried so by then end there should be no
forwarded_estimations left.

* Add tracing debug statements

Following a query through our orderbook might get difficult if the
SanitizedPriceEstimator changes things up under your nose.

To know what the SanitizedPriceEstimator does I added 3 debug log
messages:
1) "added cost of unwrapping WETH to price estimation"
   (when it modifies the winning price estimation)

2) "estimate price for native token instead of ETH"
   (when it modifies the Query in case someone wants to buy ETH)

3) "generate trivial price estimation"
   (when buy and sell token are identical)

I noticed that the 3rd message will never show because the orderbook API
returns an error in that case. But let's leave it in case this changes
in the future?

* Prevent selling ETH

Because we don't allow ETH to be the sell_token the
SanitizedPriceEstimator needs to take care of that case as well.

Surprising to me: Judging by existing e2e tests the expected behavior
it not an UnsupportedToken error but a NoLiquidity error.

* Document how SanitizedPriceEstimator::estimates works

In order to optimize the performance we need to use
PriceEstimating::estimates of the inner price estimator which estimates
a batch of queries.
This lead to some confusing code which doesn't make it obvious how
trivial and forwarded results get merged back together.
Previously I tried splitting the code into smaller functions and ginving
them descriptive names but that wasn't good enough.
Instead there is only one function again with exhaustive comments.

Co-authored-by: Martin Beckmann <martin@Martins-MacBook-Pro.fritz.box>