Fix API metrics (gnosis/gp-v2-services#1562)
Fixes over-reporting of API metrics for non-200 requests.

This is a reworked version of #1475.

I admit this was a huge waste of time. I've learned that warp is an over-complicated library that lets you do a lot of complicated stuff that you don't need, but fails at very simple tasks. I'll never use it in my projects, and will advocate against it in the future. I am now certain that I prefer a more classic approach, with routers and middlewares.

### Test Plan

Existing unit tests, manual testing:

- check that 200s are reported correctly for multiple API endpoints,
- check that 400s and 404s are reported correctly for multiple API endpoints,
- check that requests that don't match any endpoint are reported correctly.

### Release notes

This PR changes names of API metrics. Be sure to check/update dashboards prior to the release.