Propose new `/auction` endpoint (gnosis/gp-v2-services#1644)
This PR adds an initial version of the `/auction` endpoint to the open API specification for the orderbook service.

This proposes what could be the new `/auction` endpoint that the driver uses for populating getting batches. Currently, it is just `v2/solvable_orders` with the reference prices, but we may decide to add more to it in the future (there was discussion of including baseline liquidity here as well to reduce the amount of work required for new solvers - allowing them to integrate with our public `orderbook` API directly without needing to also run a `driver`).

One thing that we should probably discuss (but not necessarily important for this PR) is whether or not auctions should be sealed (i.e. multiple queries of `/auction` within a small time frame return the exact same data) or continuous (a new auction on each block). For canonical batches - it makes more sense to have them "sealed", but in the general case, I'm slightly leaning towards having them continuous, although this makes solution competition a bit more difficult if we decide to not have it happen in the driver.

### Test Plan

CI - we validate OpenAPI file 🎉.

Added an empty route  always returning 500 - you can try it out with:
```
% curl -s http://localhost:8080/api/v1/auction
{"errorType":"InternalServerError","description":""}
```

And check that the error metrics for it increased!:
```
% curl -s http://localhost:9586/metrics | rg 'gp_v2_api_api_requests_complete' | rg 'v1/auction'
gp_v2_api_api_requests_complete{method="v1/auction",status_code="500"} 1
```