Trade liquidity orders at limit price (gnosis/gp-v2-services#1544)
This PR introduces the logic for trading liquidity orders at order limit prices. In order to do so, I am modifying the settlement struct to also store non-uniform clearing price information. These prices are stored by using the sellPrice from the uniform clearing prices, and a separately stored buyPrice. This pair of (sellPrice, buyPrice) will then be similar to the limit order price of the liquidity order modulo rounding.
This new information gets encoded in the following way for the smart contract: I am replacing the [uniform_clearing_price_vec] with [[uniform_clearing_price_vec], [buy_tokens_of_liquidity_orders]]
and then all I need to do is to give the liquidity orders an index referring to the second part of the price vector.

Test Plan
This is the settlement where Martin(0xe63a) only needs to send 11688 stable tokens for his buy order to our settlement contract, as he gets the uniform clearing price.
With logic from this PR, we generate the following settlement, Martin has to pay 11725. Observe the correctness of the solution.