Add case in sanitizing price estimator for WETH -> ETH (gnosis/gp-v2-services#1656)
When implementing #1649 I noticed that native ETH orders aren't working as expected because estimating native token price for `0xeee...eee` are failing. I started looking at how to fix it and noticed that there is an insidious way to get queries where `sell_token == buy_token` by trying to do a price query from `WETH -> ETH`.

Currently, this will never happen because:
- The FE will automatically make unwrap transactions if you try to trade WETH -> ETH
- The `EthAwareMinFeeCalculator` is already swapping normalizing `buy_token == BUY_ETH_ADDRESS` to `buy_token = native_token` for us, so it would always reach the price estimators as a `sell_token: WETH, buy_token: WETH` query anyway. (**NOTE**: This means that the price estimate doesn't seem to be adding extra gas for the WETH unwrap in the price estimates. I will fix this in a follow up).

Note that this still doesn't fix my original problem, but that is a larger fix that happen in pieces (including this one 😄).

### Test Plan

Added this particular case to the existing unit test.